athena/evaluation/pyproject.toml
[tool.poetry]
name = "evaluation"
version = "0.1.0"
description = ""
authors = ["Dominik Remo <47261058+DominikRemo@users.noreply.github.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "3.12.*"
pandas = "2.2.3"
matplotlib = "3.10.0"
langid = "1.1.6"
mysql-connector-python = "9.1.0"
python-dotenv = "1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
